Unbound: integrate endpoints, but leave model in place to avoid migration issues and force de-install of the plugin.
