+++ src/opnsense/mvc/app/models/OPNsense/Unbound/ACL/ACL.xml
@@ -38,4 +38,11 @@
             <pattern>api/diagnostics/log/core/resolver/*</pattern>
         </patterns>
     </page-diagnostics-logs-resolver>
+    <page-services-unbound>
+        <name>Services: Unbound (mvc)</name>
+        <patterns>
+            <pattern>ui/unbound/*</pattern>
+            <pattern>api/unbound/*</pattern>
+        </patterns>
+    </page-services-unbound>
 </acl>
