+++ src/opnsense/mvc/app/models/OPNsense/Unbound/Menu/Menu.xml
@@ -12,6 +12,8 @@
             </ACL>
             <Statistics order="50" url="/ui/unbound/stats"/>
             <LogFile VisibleName="Log File" order="100" url="/ui/diagnostics/log/core/resolver"/>
+            <Blacklist order="50" url="/ui/unbound/dnsbl/index"/>
+            <Miscellaneous order="60" url="/ui/unbound/miscellaneous/index"/>
         </Unbound>
     </Services>
 </menu>
