--- src/opnsense/mvc/app/models/OPNsense/Unboundplus/ACL/ACL.xml
@@ -1,9 +1,0 @@
-<acl>
-   <page-services-unbound>
-      <name>Services: Unbound DNSBL</name>
-      <patterns>
-         <pattern>ui/unboundplus/*</pattern>
-         <pattern>api/unboundplus/*</pattern>
-      </patterns>
-   </page-services-unbound>
-</acl>
