+++ src/opnsense/mvc/app/views/OPNsense/Unbound/miscellaneous.volt
@@ -1,5 +1,5 @@
 {#
- # Copyright (c) 2019 Deciso B.V.
+ # Copyright (c) 2019-2020 Deciso B.V.
  # Copyright (c) 2019 Michael Muenz <m.muenz@gmail.com>
  # All rights reserved.
  #
@@ -25,30 +25,34 @@
  # POSSIBILITY OF SUCH DAMAGE.
  #}
 
+ <script>
+     $(document).ready(function() {
+         var data_get_map = {'frm_miscellaneous_settings':"/api/unbound/miscellaneous/get"};
+         mapDataToFormUI(data_get_map).done(function(data){
+             formatTokenizersUI();
+             $('.selectpicker').selectpicker('refresh');
+         });
+         $("#saveAct").SimpleActionButton({
+            onPreAction: function() {
+                const dfObj = new $.Deferred();
+                saveFormToEndpoint("/api/unbound/miscellaneous/set", 'frm_miscellaneous_settings', function(){
+                    dfObj.resolve();
+                });
+                return dfObj;
+            }
+        });
+     });
+ </script>
+
 <div class="content-box" style="padding-bottom: 1.5em;">
-    {{ partial("layout_partials/base_form",['fields':dnsblForm,'id':'frm_dnsbl_settings'])}}
+    {{ partial("layout_partials/base_form",['fields':miscellaneousForm,'id':'frm_miscellaneous_settings'])}}
     <div class="col-md-12">
         <hr />
-        <button class="btn btn-primary" id="saveAct" type="button"><b>{{ lang._('Save') }}</b> <i id="saveAct_progress"></i></button>
+        <button class="btn btn-primary" id="saveAct"
+                data-endpoint='/api/unbound/service/reload'
+                data-label="{{ lang._('Save') }}"
+                data-error-title="{{ lang._('Error reloading unbound') }}"
+                type="button">
+        </button>
     </div>
 </div>
-
-<script>
-    $(function() {
-        var data_get_map = {'frm_dnsbl_settings':"/api/unboundplus/dnsbl/get"};
-        mapDataToFormUI(data_get_map).done(function(data){
-            formatTokenizersUI();
-            $('.selectpicker').selectpicker('refresh');
-        });
-
-        // link save button to API set action
-        $("#saveAct").click(function(){
-            saveFormToEndpoint(url="/api/unboundplus/dnsbl/set", formid='frm_dnsbl_settings',callback_ok=function(){
-                $("#saveAct_progress").addClass("fa fa-spinner fa-pulse");
-                ajaxCall(url="/api/unboundplus/service/dnsbl", sendData={}, callback=function(data,status) {
-                    $("#saveAct_progress").removeClass("fa fa-spinner fa-pulse");
-                });
-            });
-        });
-    });
-</script>
