+++ src/opnsense/service/conf/actions.d/actions_unbound.conf
@@ -51,3 +51,10 @@ command:/usr/local/bin/flock -n -E 0 -o /tmp/unbound_start.lock /usr/local/opnse
 parameters:
 type:script
 message:Start Unbound
+
+[dnsbl]
+command:/usr/local/opnsense/scripts/unbound/dnsbl.py
+parameters:
+type:script
+message:fetching and applying DNSBLs
+description: Download Unbound DNSBLs and restart
