+++ src/opnsense/service/conf/actions.d/actions_unboundplus.conf
@@ -1,5 +1,9 @@
+###################################################################
+# Compatibility layer, to be removed in a future version
+# - keeps old plugin cron jobs functional
+###################################################################
 [dnsbl]
-command:/usr/local/opnsense/scripts/OPNsense/Unboundplus/dnsbl.py
+command:/usr/local/opnsense/scripts/unbound/dnsbl.py
 parameters:
 type:script
 message:fetching and applying DNSBLs
