+++ src/opnsense/service/templates/OPNsense/Unbound/core/miscellaneous.conf
