--- src/opnsense/service/templates/OPNsense/Unboundplus/+TARGETS
@@ -1,5 +1,0 @@
-dnsbl.inc:/var/unbound/etc/dnsbl.inc
-whitelist.inc:/var/unbound/etc/whitelist.inc
-miscellaneous.conf:/var/unbound/etc/miscellaneous.conf
-dot.conf:/var/unbound/etc/dot.conf
-lists.inc:/var/unbound/etc/lists.inc
